ch(move-test-file): move test file to tests folder
[Finishes 159987617]
