+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import logo from './logo.svg';
-import './App.css';
+import './styles/App.css';
 
 const App = () => (
     <div className="App">
